// packages/primitives/src/components/Settings/SettingsSection.svelte.d.ts
import type { Snippet } from 'svelte';
interface Props {
    title: string;
    description?: string;
    icon?: Snippet;
    children: Snippet;
}
declare const SettingsSection: import("svelte").Component<Props, {}, "">;
type SettingsSection = ReturnType<typeof SettingsSection>;
export default SettingsSection;
//# sourceMappingURL=SettingsSection.svelte.d.ts.map