// packages/primitives/src/components/Settings/SettingsGroup.svelte.d.ts
import type { Snippet } from 'svelte';
interface Props {
    label?: string;
    children: Snippet;
    orientation?: 'vertical' | 'horizontal';
}
declare const SettingsGroup: import("svelte").Component<Props, {}, "">;
type SettingsGroup = ReturnType<typeof SettingsGroup>;
export default SettingsGroup;
//# sourceMappingURL=SettingsGroup.svelte.d.ts.map