// packages/primitives/src/components/Settings/SettingsField.svelte.d.ts
import type { Snippet } from 'svelte';
interface Props {
    label: string;
    description?: string;
    children: Snippet;
    class?: string;
}
declare const SettingsField: import("svelte").Component<Props, {}, "">;
type SettingsField = ReturnType<typeof SettingsField>;
export default SettingsField;
//# sourceMappingURL=SettingsField.svelte.d.ts.map