// packages/primitives/src/components/Settings/SettingsToggle.svelte.d.ts
interface Props {
    label: string;
    description?: string;
    value: boolean;
    disabled?: boolean;
}
declare const SettingsToggle: import("svelte").Component<Props, {}, "value">;
type SettingsToggle = ReturnType<typeof SettingsToggle>;
export default SettingsToggle;
//# sourceMappingURL=SettingsToggle.svelte.d.ts.map